Home-FEAT-Call home page on backend (error)
Currently I get a 401 Unauthorized error as I'm not saving the tokens
and passing them back so the backend doesn't know if I'm
authenticated.

This works if you take out the 'before_action :authenticate_user!' in
the backend controller as you don't need to sign in to access the page then.